2016/17: both parts
Part I is quite easy. In order to save resources I iterate each possible path one step at a time. As soon as I find the vault, I can drop all other computations.

This means that for the second path I need to calculate all possible paths — because the longest one will be the last one to be discovered by definition. But come to think about it: I don’t think there is a way to tell how long a path can be beforehand, so I think I need to calculate all of them either way.

It takes 5-10 sec to find solution to part II and I will take it.